New feature: adđing & save image into database
So i complete it: admin can now upload tour's image to show it on the screen.
yeah, encountered a tricky problem when changing the .css file: cache css
so now you can choose an image to upload, the server automatically resize it to 900px height, convert to a bit array and store it into the database. when client go to page, we load the bit array back then use base64 convert format to get it back to image format.